requirements_test.txt
-r requirements.txt

# Check code coverage
coverage==4.2

# Mock functions for testing
mock>=1.0.1

# Check style of code
flake8>=2.1.0

# Run tests for different Python versions
tox>=1.7.0

# Utility for working with environment variables
django-environ>=0.4.0

# For a nicer testing setup
pytest

# For testing Django with pytest
pytest-django

# For mocking different functions
pytest-mock
